Refactor Examples, updated readme
