+++ examples/typechat.examplesLib/ITextRequestRouter.cs
