+++ examples/typechat.examplesLib/VectorTextIndex.cs
